+++ click_odoo_contrib/restoredb.py
@@ -4,6 +4,7 @@
 
 import os
 import shutil
+import subprocess
 
 import click
 import click_odoo
@@ -14,7 +15,7 @@ from ._dbutils import db_exists, db_management_enabled, reset_config_parameters
 from .backupdb import DBDUMP_FILENAME, FILESTORE_DIRNAME, MANIFEST_FILENAME
 
 
-def _restore_from_folder(dbname, backup, copy=True, jobs=1):
+def _restore_from_folder(dbname, backup, copy=True, jobs=1, neutralize=False):
     manifest_file_path = os.path.join(backup, MANIFEST_FILENAME)
     dbdump_file_path = os.path.join(backup, DBDUMP_FILENAME)
     filestore_dir_path = os.path.join(backup, FILESTORE_DIRNAME)
@@ -29,12 +30,21 @@ def _restore_from_folder(dbname, backup, copy=True, jobs=1):
 
     odoo.service.db._create_empty_database(dbname)
     pg_args = ["--jobs", str(jobs), "--dbname", dbname, "--no-owner", dbdump_file_path]
-    if odoo.tools.exec_pg_command("pg_restore", *pg_args):
+    pg_env = odoo.tools.misc.exec_pg_environ()
+    r = subprocess.run(
+        ["pg_restore", *pg_args],
+        env=pg_env,
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.STDOUT,
+    )
+    if r.returncode != 0:
         raise click.ClickException("Couldn't restore database")
     if copy:
         # if it's a copy of a database, force generation of a new dbuuid
         reset_config_parameters(dbname)
     with OdooEnvironment(dbname) as env:
+        if neutralize and odoo.release.version_info >= (16, 0):
+            odoo.modules.neutralize.neutralize_database(env.cr)
         if os.path.exists(filestore_dir_path):
             filestore_dest = env["ir.attachment"]._filestore()
             shutil.copytree(filestore_dir_path, filestore_dest)
@@ -48,9 +58,12 @@ def _restore_from_folder(dbname, backup, copy=True, jobs=1):
     odoo.sql_db.close_db(dbname)
 
 
-def _restore_from_file(dbname, backup, copy=True):
+def _restore_from_file(dbname, backup, copy=True, neutralize=False):
     with db_management_enabled():
-        odoo.service.db.restore_db(dbname, backup, copy)
+        extra_kwargs = {}
+        if odoo.release.version_info >= (16, 0):
+            extra_kwargs["neutralize_database"] = neutralize
+        odoo.service.db.restore_db(dbname, backup, copy, **extra_kwargs)
         odoo.sql_db.close_db(dbname)
 
 
@@ -78,6 +91,16 @@ def _restore_from_file(dbname, backup, copy=True):
     ),
 )
 @click.option(
+    "--neutralize",
+    is_flag=True,
+    show_default=True,
+    help=(
+        "Neutralize the database after restore. This will disable scheduled actions, "
+        "outgoing emails, and sets other external providers in test mode. "
+        "This works only in odoo 16.0 and above."
+    ),
+)
+@click.option(
     "--jobs",
     help=(
         "Uses this many parallel jobs to restore. Only used to "
@@ -94,7 +117,7 @@ def _restore_from_file(dbname, backup, copy=True):
         exists=True, file_okay=True, dir_okay=True, readable=True, resolve_path=True
     ),
 )
-def main(env, dbname, source, copy, force, jobs):
+def main(env, dbname, source, copy, force, neutralize, jobs):
     """Restore an Odoo database backup.
 
     This script allows you to restore databses created by using the Odoo
@@ -110,7 +133,11 @@ def main(env, dbname, source, copy, force, jobs):
         click.echo(click.style(msg, fg="yellow"))
         with db_management_enabled():
             odoo.service.db.exp_drop(dbname)
+    if neutralize and odoo.release.version_info < (16, 0):
+        raise click.ClickException(
+            "--neutralize option is only available in odoo 16.0 and above"
+        )
     if os.path.isfile(source):
-        _restore_from_file(dbname, source, copy)
+        _restore_from_file(dbname, source, copy, neutralize)
     else:
-        _restore_from_folder(dbname, source, copy, jobs)
+        _restore_from_folder(dbname, source, copy, jobs, neutralize)