+++ tests/test_restoredb.py
@@ -26,6 +26,7 @@ def _createdb(dbname):
 
 
 def _dropdb(dbname):
+    odoo.sql_db.close_all()
     subprocess.check_call(["dropdb", "--if-exists", dbname])
 
 
@@ -105,5 +106,29 @@ def test_db_restore_move(backup):
         assert db_exists(TEST_DBNAME)
         # when database is moved, default params are preserved
         _check_default_params(TEST_DBNAME, original_db, operator.eq)
+    finally:
+        _dropdb_odoo(TEST_DBNAME)
+
+
+def test_db_restore_neutralize(backup):
+    assert not db_exists(TEST_DBNAME)
+    backup_path, _original_db = backup
+    try:
+        result = CliRunner().invoke(
+            restoredb, ["--neutralize", TEST_DBNAME, backup_path]
+        )
+        if odoo.release.version_info < (16, 0):
+            assert result.exit_code != 0, result.output
+            assert (
+                "--neutralize option is only available in odoo 16.0 and above"
+                in result.output
+            )
+            assert not db_exists(TEST_DBNAME)
+        else:
+            assert result.exit_code == 0
+            assert db_exists(TEST_DBNAME)
+            with click_odoo.OdooEnvironment(database=TEST_DBNAME) as env:
+                IrConfigParameters = env["ir.config_parameter"]
+                assert IrConfigParameters.get_param("database.is_neutralized") == "true"
     finally:
         _dropdb_odoo(TEST_DBNAME)